pyproject.toml
[project]
name = "pydantic_ai_demo"
version = "0.1.0"
description = "A simple demo of an agent with Instructor (Pydantic AI) and Logfire."
authors = [{ name = "Cascade AI", email = "cascade@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "instructor>=1.0.0",
    "logfire>=0.20.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pydantic-agent-demo = "pydantic_ai_demo.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# UV specific configurations can go here if needed
